maj de graphic.hpp

+++ header/Graphic_Fonction.hpp
@@ -12,7 +12,7 @@ std::string Fenetre_acceuil(sdl2::window* fenêtre, int* dim_x, int* dim_y);
 int Choix_Niveau(sdl2::window* fenêtre);
 void Centrale_tab(sdl2::window* fenêtre);
 void Controle_win(int level);
-void Commande(sdl2::window * fenêtre, int *dim_x, double *dim_y);
+void Commande(sdl2::window * fenêtre, int *dim_x, int *dim_y);
 void Niveau1(sdl2::window* fenêtre);
 void SalleDeControle();
 void Radioprotection_win(int level);
